Version1_jammu_app/requirements.txt
flask==2.3.2
torch==2.0.1
torchvision==0.15.2
transformers==4.31.0
pillow==9.5.0
opencv-python==4.7.0.72
numpy==1.24.3
tqdm==4.65.0